performance/requirements.txt
scikit-image
optax
jaxopt
flax
datasets
tqdm
numpy
einops
pyseltongue
pycryptodome
